Basta gi minusan ra nakong mga todo list nimo

+++ src/Entities/Entity.java
@@ -3,12 +3,11 @@ package Entities;
 public abstract class Entity {
     protected double hp;
     protected double maxHp;
-    protected int attack;
-    protected int maxAttack;
-    protected int defense;
-    protected int maxDefense;
+    protected double attack;
+    protected double maxAttack;
+    protected double defense;
+    protected double maxDefense;
     protected double dmgResistance;
-    protected int speed;
+    protected double speed;
     protected double accuracy;
-
 }
